moves mock package offer into utils folder
