+++ src/utils/mock-package-offer.js
@@ -1,0 +1,156 @@
+const mockPackageOffer = {
+  'description': 'Lorem ipsum dolor sit amet, magna est placerat velit auctor aenean sodales. Interdum metus, mollis lorem tempor vel amet quisque hymenaeos, luctus in vehicula tristique sed molestie, sodales gravida vivamus placerat. Aliquam aliquet tempus, suspendisse dolor aenean rhoncus cras a, in turpis curabitur semper mi in mattis, justo augue nisl sit magna urna, nonummy proin id. Ac vel nisl pellentesque non, non arcu justo, dictum tempor aenean lacus ipsum. Nullam ut montes vitae suspendisse sociis vitae. Quam a ante tristique etiam tellus, morbi vitae eget eget in. Tortor mi nulla hac, faucibus cras in nascetur amet maecenas, per nulla vitae in et.',
+  'hotel': {
+    'id': 'hotel:mhid.ygky5mL',
+    'name': 'Viva Sunrise',
+    'images': [
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_668866.jpg'
+      },
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_668888.jpg'
+      },
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_773200.jpg'
+      },
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_227930.jpg'
+      },
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_668879.jpg'
+      },
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_668900.jpg'
+      },
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_773205.jpg'
+      },
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_773201.jpg'
+      },
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_668904.jpg'
+      },
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_773206.jpg'
+      },
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_668902.jpg'
+      },
+      {
+        'type': 'image/jpeg',
+        'displaySequence': null,
+        'primary': null,
+        'uri': '/H0/00/36/37/TCUK_STEP51_v1_546203.jpg'
+      }
+    ],
+    'starRating': '3',
+    'place': {
+      'name': 'Port d\'Alcúdia',
+      'country': 'Spain',
+      'region': 'Balearic Islands'
+    }
+  },
+  'flights': {
+    'outbound': [{
+      'number': 'TCX1124',
+      'departure': {
+        'localDateTime': '2016-06-13T06:10',
+        'airport': {
+          'code': 'LGW',
+          'name': 'Gatwick'
+        }
+      },
+      'arrival': {
+        'localDateTime': '2016-06-13T09:30',
+        'airport': {
+          'code': 'PMI'
+        }
+      },
+      'carrier': {
+        'code': 'TCX'
+      }
+    }],
+    'inbound': [{
+      'number': 'TCX1125',
+      'departure': {
+        'localDateTime': '2016-06-20T10:30',
+        'airport': {
+          'code': 'PMI',
+          'name': 'Son Sant Joan'
+        }
+      },
+      'arrival': {
+        'localDateTime': '2016-06-20T11:55',
+        'airport': {
+          'code': 'LGW'
+        }
+      },
+      'carrier': {
+        'code': 'TCX'
+      }
+    }]
+  },
+  'price': {
+    'total': '3076',
+    'perPerson': '1537.98',
+    'currency': 'GBP'
+  },
+  'provider': {
+    'id': 'ac-searcher',
+    'reference': 'A01A37'
+  },
+  'nights': 7,
+  'amenities': {
+    'outdoorpool': '3 stk.',
+    'distancetobeach': '0 m',
+    'distancetocenter': '500 m',
+    'bar': true,
+    'childrenpool': true,
+    'elevator': true,
+    'poolbar': true,
+    'restaurant': true,
+    'minimarket': true,
+    'cleaningdaysperweek': '5',
+    'wifi': true,
+    'waterslide': true,
+    'lolloandbernie': false,
+    'isadulthotel': false,
+    'allinclusive': false
+  }
+};
+
+export default mockPackageOffer;
